+++ internal/back/handler/handler.go
@@ -1,11 +1,11 @@
 package handler
 
-import "github.com/Alekseyt9/upscaler/internal/back/services/s3stor"
+import "github.com/Alekseyt9/upscaler/internal/back/services/s3store"
 
 type FrontHandler struct {
-	s3 s3stor.S3Store
+	s3 s3store.S3Store
 }
 
-func New(s3 s3stor.S3Store) *FrontHandler {
+func New(s3 s3store.S3Store) *FrontHandler {
 	return &FrontHandler{s3: s3}
 }
