+++ internal/back/services/s3store/custom_endpoint_resolver.go
@@ -1,4 +1,4 @@
-package s3stor
+package s3store
 
 import (
 	"context"
